requirements.txt

fastapi==0.95.0
uvicorn[standard]==0.22.0
websockets==11.0.3
python-multipart==0.0.6
pydub==0.25.1
numpy==1.26.0
soundfile==0.12.1
torch==2.2.0
torchaudio==2.2.0
transformers==4.40.0
pyannote.audio==2.1.1
librosa==0.10.0
scipy==1.11.3
webrtcvad==2.0.10
aiofiles==23.1.0
typing_extensions
pydantic
sentencepiece
faster-whisper==0.6.0
sounddevice==0.4.6
jiwer==3.0.5
